Update quarantine-admin-manage-messages-files.md
There is no 'Quarantine Administrator' role I can find documented for Exchange Online RBAC. That role only is documented as part of the Defender 365 so-called Unified RBAC.

The appropriate role in Exchange Online RBAC appears to be "Hygiene Management".  Based on what I've read, both that role AND the "Quarantine Administrator" role are needed.